Add custom basic auth. change check ids in requests

+++ database/migrations/2023_02_20_165303_create_locations_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('locations', function (Blueprint $table) {
             $table->id();
-            $table->double('latitude');
-            $table->double('longitude');
+            $table->decimal('latitude', 20, 17);
+            $table->decimal('longitude', 20, 17);
         });
     }
